This commit was manufactured by cvs2svn to create tag 'WX_2_2_6'.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/tags/WX_2_2_6@9545 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/motif/dcclient.cpp
@@ -323,7 +323,7 @@ void wxWindowDC::DoDrawArc( wxCoord x1, wxCoord y1, wxCoord x2, wxCoord y2, wxCo
 
     wxCoord dx = xx1 - xxc;
     wxCoord dy = yy1 - yyc;
-    double radius = sqrt (dx * dx + dy * dy);
+    double radius = sqrt ((double)(dx * dx + dy * dy));
     wxCoord r = (wxCoord) radius;
 
     double radius1, radius2;
